Fix testcases
