+++ faebryk/exporters/netlist/__init__.py
@@ -2,4 +2,5 @@
 # SPDX-License-Identifier: MIT
 
 import faebryk.exporters.netlist.kicad
+import faebryk.exporters.netlist.graph
 from .netlist import *
