+++ samples/experiment2.py
@@ -1,96 +1,10 @@
 # This file is part of the faebryk project
 # SPDX-License-Identifier: MIT
 
-# Temporary -------------------------------------------------------------------
-def make_t1_netlist_from_graph(comps):
-    t1_netlist = [comp.get_comp() for comp in comps]
-
-    return t1_netlist
-
-def make_graph_from_components(components):
-    import faebryk as fy
-    from faebryk.library.core import Component
-    from faebryk.libs.exceptions import FaebrykException
-    from faebryk.library.traits.component import has_footprint, has_type_description, has_footprint_pinmap
-    from faebryk.library.traits.interface import is_part_of_component
-    from faebryk.library.traits.footprint import has_kicad_footprint
-
-    class wrapper():
-        def __init__(self, component: Component) -> None:
-            self.component = component
-            self._setup_non_rec()
-
-        def _setup_non_rec(self):
-            import random
-            c = self.component
-            self.real = c.has_trait(has_footprint) and c.has_trait(has_footprint_pinmap)
-            self.properties = {}
-            self.neighbors = {}
-            if self.real:
-                self.value = c.get_trait(has_type_description).get_type_description()
-                self.properties["footprint"] = \
-                    c.get_trait(has_footprint).get_footprint().get_trait(
-                        has_kicad_footprint).get_kicad_footprint()
-            self.name = "COMP[{}:{}]@{:08X}".format(type(self.component).__name__, self.value if self.real else "virt", int(random.random()*2**32))
-            self._comp = {}
-            self._update_comp()
-
-        def _update_comp(self):
-            self._comp.update({
-                "name": self.name,
-                "real": self.real,
-                "properties": self.properties,
-                "neighbors": self.neighbors
-            })
-            if self.real:
-                self._comp["value"] = self.value
-
-        def _get_comp(self):
-            return self._comp
-
-        def get_comp(self):
-            # only executed once
-            neighbors = {}
-            for pin, interface in self.component.get_trait(has_footprint_pinmap).get_pin_map().items():
-                neighbors[pin] = []
-                for target_interface in interface.connections:
-                    if target_interface.has_trait(is_part_of_component):
-                        target_component = target_interface.get_trait(is_part_of_component).get_component()
-                        target_pinmap = target_component.get_trait(has_footprint_pinmap).get_pin_map()
-                        target_pin = list(target_pinmap.items())[list(target_pinmap.values()).index(target_interface)][0]
-                        try:
-                            target_wrapped = [i for i in wrapped_list if i.component == target_component][0]
-                        except IndexError:
-                            raise FaebrykException("Discovered associated component not in component list:", target_component)
-
-                        neighbors[pin].append({
-                          "vertex": target_wrapped._get_comp(),
-                          "pin": target_pin
-                        })
-                    else:
-                        print("Warning: {comp} pin {pin} is connected to interface without component".format(
-                            comp=self.name,
-                            #intf=target_interface,
-                            pin=pin,
-                        ))
-
-            self.neighbors = neighbors
-            self._update_comp()
-
-            return self._get_comp()
-
-    wrapped_list = list(map(wrapper, components))
-    for i in wrapped_list:
-        i.wrapped_list = wrapped_list
-
-    print("Making graph from components:\n\t{}".format("\n\t".join(map(str, components))))
-
-    return wrapped_list
-# -----------------------------------------------------------------------------
-
 def run_experiment():
     from faebryk.exporters.netlist.kicad.netlist_kicad import from_faebryk_t2_netlist
     from faebryk.exporters.netlist import make_t2_netlist_from_t1
+    from faebryk.exporters.netlist.graph import make_graph_from_components, make_t1_netlist_from_graph
     from faebryk.library.core import Component, Footprint
     from faebryk.library.library.components import CD4011, LED, NAND, Resistor, Switch
     from faebryk.library.library.footprints import DIP, SMDTwoPin
